src/beat/requirements.txt
pyotp==2.2.6
python-dotenv==0.9.1
